+++ Sources/TableViewModel.swift
@@ -95,16 +95,20 @@ extension TableCellViewModel {
 public protocol TableViewCellModelEditActions {
 
     /// The edit actions for this cell when swiping from the leading direction.
-    @available(iOS 11.0, *)
     var leadingSwipeActionConfiguration: UISwipeActionsConfiguration? { get }
 
     /// The edit actions for this cell when swiping from the trailing direction.
-    @available(iOS 11.0, *)
     var trailingSwipeActionConfiguration: UISwipeActionsConfiguration? { get }
+}
 
-    /// The row edit actions for the cell.
-    /// - Warning: `UITableViewRowAction` is deprecated in iOS 13. Implement `leadingSwipeActionConfiguration` or `trailingSwipeActionConfiguration` instead.
-    var editActions: [UITableViewRowAction] { get }
+/// Default implementation for `TableViewCellModelEditActions`.
+extension TableViewCellModelEditActions {
+
+    /// Default implementation, returns `nil`.
+    var leadingSwipeActionConfiguration: UISwipeActionsConfiguration? { return nil }
+
+    /// Default implementation, returns `nil`.
+    var trailingSwipeActionConfiguration: UISwipeActionsConfiguration? { return nil }
 }
 
 /// Protocol that needs to be implemented by custom header